# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Cyber Threat Intelligence Dashboard API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ThreatEventSeverity =
  (typeof ThreatEventSeverity)[keyof typeof ThreatEventSeverity];

export const ThreatEventSeverity = {
  critical: "critical",
  high: "high",
  medium: "medium",
  low: "low",
} as const;

export interface ThreatEvent {
  id: string;
  timestamp: string;
  attackerCountry: string;
  attackerCountryCode: string;
  attackerIp?: string;
  attackerLat: number;
  attackerLng: number;
  targetCountry: string;
  targetCountryCode: string;
  targetIp?: string;
  targetLat: number;
  targetLng: number;
  attackType: string;
  severity: ThreatEventSeverity;
  port?: number;
  protocol?: string;
  description?: string;
  technique?: string;
  mitreTactic?: string;
}

export interface ThreatStats {
  totalAttacks24h: number;
  criticalAttacks: number;
  highAttacks: number;
  mediumAttacks: number;
  lowAttacks: number;
  attacksPerMinute: number;
  uniqueAttackers: number;
  uniqueTargets: number;
  mostCommonAttackType: string;
  totalCountriesAffected: number;
}

export interface CountryAttackStats {
  country: string;
  countryCode: string;
  attackCount: number;
  percentage: number;
  topAttackTypes: string[];
}

export type AttackTypeStatsSeverityBreakdown = {
  critical: number;
  high: number;
  medium: number;
  low: number;
};

export interface AttackTypeStats {
  attackType: string;
  count: number;
  percentage: number;
  severityBreakdown: AttackTypeStatsSeverityBreakdown;
}

export interface TimelinePoint {
  hour: string;
  count: number;
  critical: number;
  high: number;
}

export type NewsArticleSeverity =
  (typeof NewsArticleSeverity)[keyof typeof NewsArticleSeverity];

export const NewsArticleSeverity = {
  critical: "critical",
  high: "high",
  medium: "medium",
  low: "low",
  info: "info",
} as const;

export interface NewsArticle {
  id: string;
  title: string;
  summary: string;
  url: string;
  source: string;
  publishedAt: string;
  category: string;
  tags: string[];
  severity?: NewsArticleSeverity;
  relatedCountries?: string[];
}

export type ThreatActorThreatLevel =
  (typeof ThreatActorThreatLevel)[keyof typeof ThreatActorThreatLevel];

export const ThreatActorThreatLevel = {
  critical: "critical",
  high: "high",
  medium: "medium",
  low: "low",
} as const;

export interface ThreatActor {
  name: string;
  alias: string[];
  originCountry: string;
  motivation: string;
  recentActivity: string;
  targetSectors: string[];
  threatLevel: ThreatActorThreatLevel;
}

export interface CveEntry {
  cveId: string;
  cvssScore: number;
  severity: string;
  description: string;
  affectedSoftware: string;
  publishedAt: string;
  exploitAvailable: boolean;
}

export interface TrendingThreats {
  threatActors: ThreatActor[];
  recentCves: CveEntry[];
  activeCampaigns: string[];
}

export type ListThreatsParams = {
  limit?: number;
  attackType?: string;
  severity?: ListThreatsSeverity;
};

export type ListThreatsSeverity =
  (typeof ListThreatsSeverity)[keyof typeof ListThreatsSeverity];

export const ListThreatsSeverity = {
  critical: "critical",
  high: "high",
  medium: "medium",
  low: "low",
} as const;

export type ListCyberNewsParams = {
  limit?: number;
  category?: string;
};