# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Cyber Threat Intelligence Dashboard API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /threats:
    get:
      operationId: listThreats
      tags: [threats]
      summary: List all active cyber threats
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: attackType
          in: query
          schema:
            type: string
        - name: severity
          in: query
          schema:
            type: string
            enum: [critical, high, medium, low]
      responses:
        "200":
          description: List of threats
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ThreatEvent"
...
  /news:
    get:
      operationId: listCyberNews
      tags: [news]
      summary: Get latest cybersecurity news articles
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: category
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of news articles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/NewsArticle"
...
components:
  schemas:
...
    ThreatEvent:
      type: object
      properties:
        id:
          type: string
        timestamp:
          type: string
          format: date-time
        attackerCountry:
          type: string
        attackerCountryCode:
          type: string
        attackerIp:
          type: string
        attackerLat:
          type: number
        attackerLng:
          type: number
        targetCountry:
          type: string
        targetCountryCode:
          type: string
        targetIp:
          type: string
        targetLat:
          type: number
        targetLng:
          type: number
        attackType:
          type: string
        severity:
          type: string
          enum: [critical, high, medium, low]
        port:
          type: integer
        protocol:
          type: string
        description:
          type: string
        technique:
          type: string
        mitreTactic:
          type: string
      required:
        - id
        - timestamp
        - attackerCountry
        - attackerCountryCode
        - attackerLat
        - attackerLng
        - targetCountry
        - targetCountryCode
        - targetLat
        - targetLng
        - attackType
        - severity
...
    NewsArticle:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        summary:
          type: string
        url:
          type: string
        source:
          type: string
        publishedAt:
          type: string
          format: date-time
        category:
          type: string
        tags:
          type: array
          items:
            type: string
        severity:
          type: string
          enum: [critical, high, medium, low, info]
        relatedCountries:
          type: array
          items:
            type: string
      required:
        - id
        - title
        - summary
        - url
        - source
        - publishedAt
        - category
        - tags
...