# lib/api-zod/src/generated/types/newsArticle.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Cyber Threat Intelligence Dashboard API
 * OpenAPI spec version: 0.1.0
 */
import type { NewsArticleSeverity } from "./newsArticleSeverity";

export interface NewsArticle {
  id: string;
  title: string;
  summary: string;
  url: string;
  source: string;
  publishedAt: Date;
  category: string;
  tags: string[];
  severity?: NewsArticleSeverity;
  relatedCountries?: string[];
}