# lib/api-zod/src/generated/types/threatEvent.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Cyber Threat Intelligence Dashboard API
 * OpenAPI spec version: 0.1.0
 */
import type { ThreatEventSeverity } from "./threatEventSeverity";

export interface ThreatEvent {
  id: string;
  timestamp: Date;
  attackerCountry: string;
  attackerCountryCode: string;
  attackerIp?: string;
  attackerLat: number;
  attackerLng: number;
  targetCountry: string;
  targetCountryCode: string;
  targetIp?: string;
  targetLat: number;
  targetLng: number;
  attackType: string;
  severity: ThreatEventSeverity;
  port?: number;
  protocol?: string;
  description?: string;
  technique?: string;
  mitreTactic?: string;
}